Remove discrim related website dependencies
These seem to be causing some problems in the case weights branch because the dev version of discrim's dependencies conflict with the dependencies needed by the case weights branch (specifically, parsnip). It doesn't add too much value and has been a problem before, so I think it is easiest to just remove it.